tests/requirements.txt
pytest>=7.0
requests>=2.31.0
nostr-sdk
cashu>=0.19.0
marshmallow<4
pytest-timeout
pytest-order
pytest-playwright
responses
pytest-asyncio
pytest-aiohttp
